+++ Cards/src/com/eg/cards/PutEvent.java
@@ -4,12 +4,12 @@ import com.badlogic.gdx.scenes.scene2d.Event;
 
 public class PutEvent extends Event{
 	
-	private final byte id;
+	private final int id;
 	
-	public PutEvent(byte id){
+	public PutEvent(int id){
 		this.id = id;
 	}
 	
-	public byte getID(){ return id; }
+	public int getID(){ return id; }
 
 }
